Move Twitter::Request into Twitter::REST namespace
