+++ lib/twitter/rest/favorites.rb
@@ -1,6 +1,5 @@
 require 'twitter/arguments'
 require 'twitter/error'
-require 'twitter/request'
 require 'twitter/rest/utils'
 require 'twitter/tweet'
 require 'twitter/user'
