+++ lib/twitter/rest/lists.rb
@@ -2,7 +2,6 @@ require 'twitter/arguments'
 require 'twitter/cursor'
 require 'twitter/error'
 require 'twitter/list'
-require 'twitter/request'
 require 'twitter/rest/utils'
 require 'twitter/tweet'
 require 'twitter/user'
