+++ lib/twitter/rest/places_and_geo.rb
@@ -1,6 +1,5 @@
 require 'twitter/geo_results'
 require 'twitter/place'
-require 'twitter/request'
 require 'twitter/rest/utils'
 
 module Twitter
