+++ lib/twitter/rest/spam_reporting.rb
@@ -1,4 +1,3 @@
-require 'twitter/request'
 require 'twitter/rest/utils'
 require 'twitter/user'
