+++ lib/twitter/rest/suggested_users.rb
@@ -1,5 +1,4 @@
 require 'twitter/arguments'
-require 'twitter/request'
 require 'twitter/rest/utils'
 require 'twitter/suggestion'
 require 'twitter/user'
