+++ lib/twitter/rest/trends.rb
@@ -1,5 +1,4 @@
 require 'twitter/place'
-require 'twitter/request'
 require 'twitter/rest/utils'
 require 'twitter/trend_results'
